Stage 7: the Google Drive corpus, and the bug it found
The corpus is archived in place rather than copied to a local subset. It is in
the list for its high-latency I/O pattern - stage 2 measured one 100 kB write
at 2.673 s through this mount - and copying it to local storage first would
remove exactly the property it exists to exercise.

It earned its place immediately. An rclone mount synthesises inode numbers from
a hash, so they exceed 2^63. borg stores the inode as a msgpack uint64 and
Python's arbitrary-precision integers never notice; borge decoded it as int64
and refused to read the archive at all:

    borge: item: inode value 16477067133423719032 does not fit in an int64

item.Item.Inode is now unsigned, and encoded as unsigned so it round-trips as
the number that went in rather than as a negative one. No other corpus could
have surfaced this - every real filesystem's inode fits comfortably in an
int64. Recorded as divergence 10.

84 entries identical in both directions afterwards, with borg check
--verify-data and borge check --verify-data over the shared repository.

Also loosens an over-precise assertion in the files cache test: it demanded
exactly four of five files come from the cache, but how many are dropped
depends on how many share the newest timestamp, which depends on the
filesystem's granularity. The property is "most of them, and fewer after an
edit", which is what it now asserts.

+++ internal/archive/create_linux.go
@@ -383,7 +383,7 @@ func (w *walker) fillMetadata(it *item.Item, abs string, st *unix.Stat_t) {
 	if st.Mode&unix.S_IFMT == unix.S_IFREG {
 		it.Size = &size
 	}
-	inode := int64(st.Ino)
+	inode := st.Ino
 	it.Inode = &inode
 
 	if !w.opts.NoXAttrs {
